pyproject.toml
[tool.poetry]
name = "circlex"
version = "0.1.0"
description = "A simple social media project for automatic posts generations"
authors = ["Aman Rawat <kirmolaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
drf-nested-routers = "^0.93.5"
shortuuid = "^1.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
